Create challenge#1
This challenge is asking you to write a program that will take in a dictionary containing name-marks pairs for a list of students. The marks for each student is given as an array of 3 integers. The program should then find the average of the marks for the student whose name is provided as query_name and return the average with 2 decimal places.

Here's a breakdown of the inputs and outputs of the program:

Input:
The first line of the input contains an integer n which represents the number of students' records.
The next n lines will contain the name of a student followed by their 3 marks separated by a space.
The final line of the input contains the name of the student to query (query_name).
Output:
The program should print the average of the marks for the student specified by query_name with 2 decimal places.